refactor: delete glchain and improve modularity
