+++ gltest/__init__.py
@@ -1,14 +1,15 @@
-from gltest.glchain import (
-    create_account,
-    create_accounts,
-    get_contract_factory,
-    get_gl_client,
-    get_accounts,
+from gltest.accounts import (
     get_default_account,
+    get_accounts,
+    create_accounts,
+    create_account,
 )
+from gltest.clients import (
+    get_gl_client,
+)
+from gltest.contracts import get_contract_factory
 
 __all__ = [
-    "find_contract_definition",
     "create_account",
     "create_accounts",
     "get_contract_factory",
