+++ gltest/contracts/__init__.py
@@ -1,0 +1,4 @@
+from .contract import Contract
+from .contract_factory import ContractFactory, get_contract_factory
+
+__all__ = ["Contract", "ContractFactory", "get_contract_factory"]
