--- gltest/glchain/__init__.py
@@ -1,16 +1,0 @@
-from .contract import Contract, ContractFactory, get_contract_factory
-from .client import get_gl_client, get_gl_provider
-from .account import create_account, get_accounts, get_default_account, create_accounts
-
-
-__all__ = [
-    "Contract",
-    "ContractFactory",
-    "get_contract_factory",
-    "create_account",
-    "create_accounts",
-    "get_accounts",
-    "get_default_account",
-    "get_gl_client",
-    "get_gl_provider",
-]
